+++ YouTunes/src/main/java/model/AlbumDao.java
@@ -2,6 +2,6 @@ package model;
 
 import beans.Album;
 
-public interface AlbumDao extends GenericDao<Album, Long>{
+public interface AlbumDao extends GenericDao<Album, Integer>{
 	}
