+++ YouTunes/src/main/java/model/ArtistDao.java
@@ -2,6 +2,6 @@ package model;
 
 import beans.Artist;
 
-public interface ArtistDao extends GenericDao<Artist, Long>{
+public interface ArtistDao extends GenericDao<Artist, Integer>{
 	}
